// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const BillStatus = {
  open: 'open',
  pending_cnab: 'pending_cnab',
  paid: 'paid',
  cancelled: 'cancelled'
} as const

export type BillStatus = (typeof BillStatus)[keyof typeof BillStatus]


export const TitleStatus = {
  PENDING: 'PENDING',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  OVERDUE: 'OVERDUE'
} as const

export type TitleStatus = (typeof TitleStatus)[keyof typeof TitleStatus]


export const TitleStakeholder = {
  PEOPLE: 'PEOPLE',
  STUDENT: 'STUDENT',
  PARTNER: 'PARTNER',
  INSTRUCTOR: 'INSTRUCTOR',
  EMPLOYEE: 'EMPLOYEE',
  COMPANY: 'COMPANY',
  NONE: 'NONE'
} as const

export type TitleStakeholder = (typeof TitleStakeholder)[keyof typeof TitleStakeholder]